data_pipeline/requirements.txt
pandas
sqlalchemy
psycopg2-binary
pyyaml
tqdm
loguru
